Extract teams to constants
